# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Student Admission & Entry Test Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: Admin - list all applications
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, rejected]
        - in: query
          name: program
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Applications list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationListResponse"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Student submits registration form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Application created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get application by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
    patch:
      operationId: updateApplication
      tags: [applications]
      summary: Update application (admin approve/reject or student draft)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpdate"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
    delete:
      operationId: deleteApplication
      tags: [applications]
      summary: Delete application
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /applications/by-student/{studentId}:
    get:
      operationId: getApplicationByStudent
      tags: [applications]
      summary: Get application by student ID
      parameters:
        - in: path
          name: studentId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"

  # Tests
...
  /interviews:
    get:
      operationId: listInterviews
      tags: [interviews]
      summary: List all interviews
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [scheduled, completed, rejected]
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Interviews list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewListResponse"
    post:
      operationId: createInterview
      tags: [interviews]
      summary: Schedule an interview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewInput"
      responses:
        "201":
          description: Interview scheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"
  /interviews/{id}:
    get:
      operationId: getInterview
      tags: [interviews]
      summary: Get interview by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Interview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"
    patch:
      operationId: updateInterview
      tags: [interviews]
      summary: Update interview notes/status/remarks
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewUpdate"
      responses:
        "200":
          description: Updated interview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"
    delete:
      operationId: deleteInterview
      tags: [interviews]
      summary: Delete interview
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Violations
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    Admin:
      type: object
      required: [id, email, name]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
    AuthResponse:
      type: object
      required: [token, admin]
      properties:
        token:
          type: string
        admin:
          $ref: "#/components/schemas/Admin"

    # Student
...
    Application:
      type: object
      required: [id, studentId, status, createdAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        rollNumber:
          type: ["string", "null"]
        fullName:
          type: ["string", "null"]
        fatherName:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        pictureUrl:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        district:
          type: ["string", "null"]
        province:
          type: ["string", "null"]
        schoolCollegeName:
          type: ["string", "null"]
        previousClass:
          type: ["string", "null"]
        board:
          type: ["string", "null"]
        passingYear:
          type: ["string", "null"]
        obtainedMarks:
          type: ["integer", "null"]
        totalMarks:
          type: ["integer", "null"]
        percentage:
          type: ["number", "null"]
        programSelection:
          type: ["string", "null"]
        campusSelection:
          type: ["string", "null"]
        guardianName:
          type: ["string", "null"]
        guardianContact:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, pending, approved, rejected]
        adminNotes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    ApplicationInput:
      type: object
      required: [studentId]
      properties:
        studentId:
          type: integer
        fullName:
          type: string
        fatherName:
          type: string
        gender:
          type: string
        dateOfBirth:
          type: string
        pictureUrl:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        address:
          type: string
        city:
          type: string
        district:
          type: string
        province:
          type: string
        schoolCollegeName:
          type: string
        previousClass:
          type: string
        board:
          type: string
        passingYear:
          type: string
        obtainedMarks:
          type: integer
        totalMarks:
          type: integer
        percentage:
          type: number
        programSelection:
          type: string
        campusSelection:
          type: string
        guardianName:
          type: string
        guardianContact:
          type: string
        status:
          type: string
          enum: [draft, pending]
    ApplicationUpdate:
      type: object
      properties:
        fullName:
          type: string
        fatherName:
          type: string
        gender:
          type: string
        dateOfBirth:
          type: string
        pictureUrl:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        address:
          type: string
        city:
          type: string
        district:
          type: string
        province:
          type: string
        schoolCollegeName:
          type: string
        previousClass:
          type: string
        board:
          type: string
        passingYear:
          type: string
        obtainedMarks:
          type: integer
        totalMarks:
          type: integer
        percentage:
          type: number
        programSelection:
          type: string
        campusSelection:
          type: string
        guardianName:
          type: string
        guardianContact:
          type: string
        status:
          type: string
          enum: [draft, pending, approved, rejected]
        adminNotes:
          type: string
    ApplicationListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Application"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # Test
...
    Interview:
      type: object
      required: [id, studentId, scheduledAt, status, createdAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        rollNumber:
          type: ["string", "null"]
        scheduledAt:
          type: string
        venue:
          type: ["string", "null"]
        status:
          type: string
          enum: [scheduled, completed, rejected]
        notes:
          type: ["string", "null"]
        remarks:
          type: ["string", "null"]
        createdAt:
          type: string
    InterviewInput:
      type: object
      required: [studentId, scheduledAt]
      properties:
        studentId:
          type: integer
        scheduledAt:
          type: string
        venue:
          type: string
        notes:
          type: string
    InterviewUpdate:
      type: object
      properties:
        scheduledAt:
          type: string
        venue:
          type: string
        status:
          type: string
          enum: [scheduled, completed, rejected]
        notes:
          type: string
        remarks:
          type: string
    InterviewListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Interview"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # Violations
...